Prevent modification of global http.DefaultClient.
In case when user of the library does not provide http.Client instance,
GWC used http.DefaultClient. This is bad since GWC replaces RoundTripper
to enable retry mechanism. This means that alternative RoundTripper was
set for all users of http.DefaultClient. With this change, new instance
of the client is created if user did not provide one.